+++ src/main/java/com/tanksDevs/system/entity/hitBox/HitBox.java
@@ -3,8 +3,8 @@ package com.tanksDevs.system.entity.hitBox;
 public interface HitBox {
 
     boolean checkCollision(HitBox other);
-    int getBottomX();
-    int getBottomY();
-    int getTopX();
-    int getTopY();
+    double getBottomX();
+    double getBottomY();
+    double getTopX();
+    double getTopY();
 }
